backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "confluent-kafka",
    "psycopg2-binary",
    "sse-starlette",
]
